backend/requirements_suggested.txt
# Tested combo for Windows CPU + Python 3.10/3.11
numpy>=1.24
opencv-python>=4.10
tensorflow==2.15.0
keras==2.15.0
cvzone>=1.6.1
mediapipe>=0.10.14
pillow>=10.3.0
pyenchant>=3.2.2
pyttsx3>=2.90


